requirements.txt
    streamlit
    langchain
    langchain-openai
    langchain-community # Includes FAISS, HuggingFaceEmbeddings (if needed as fallback)
    langchain-core
    langgraph
    langchain-tavily
    faiss-cpu # Using CPU version for broader compatibility
    # sentence-transformers # No longer primary embedding, keep if needed as fallback
    # torch # No longer primary embedding
    # transformers # No longer primary embedding
    pypdf2
    openai>=1.0 # Ensure openai library is compatible with langchain-openai
    tiktoken
    python-dotenv # For loading .env file
    pandas # Added for Excel processing
    openpyxl # Added for Excel processing
    requests # For making HTTP requests to external APIs
    # Add any other dependencies you need for your project here